検索画面 MVCにリファクタ
